Rename "AgeCard" to "PatientsByAgeCard"
