+++ pages/cards/_card.vue
@@ -12,7 +12,9 @@
     <inspections-transition-card
       v-else-if="this.$route.params.card == 'number-of-tested-cases'"
     />
-    <age-card v-else-if="this.$route.params.card == 'patients-by-age'" />
+    <patients-by-age-card
+      v-else-if="this.$route.params.card == 'patients-by-age'"
+    />
     <sickbeds-used-rate-card
       v-else-if="this.$route.params.card == 'patietns-and-sickedbeds'"
     />
@@ -30,7 +32,7 @@ import PatientsBreakdownCard from '@/components/cards/PatientsBreakdownCard.vue'
 import PatientsTransitionCard from '@/components/cards/PatientsTransitionCard.vue'
 import PatientsDetailCard from '@/components/cards/PatientsDetailCard.vue'
 import InspectionsTransitionCard from '@/components/cards/InspectionsTransitionCard.vue'
-import AgeCard from '@/components/cards/AgeCard.vue'
+import PatientsByAgeCard from '@/components/cards/PatientsByAgeCard.vue'
 import SickbedsUsedRateCard from '@/components/cards/SickbedsUsedRateCard.vue'
 import ConsultationsTransitionCard from '@/components/cards/ConsultationsTransitionCard.vue'
 
@@ -40,7 +42,7 @@ export default {
     PatientsTransitionCard,
     PatientsDetailCard,
     InspectionsTransitionCard,
-    AgeCard,
+    PatientsByAgeCard,
     SickbedsUsedRateCard,
     ConsultationsTransitionCard
   },
